// src/app/sso/model/catalogoUsuarios.ts
export interface UsuarioDTO {
    usuario: string;
    usrPWD: string;
    isUsuario: string;
    nombre: string;
    apellidoPaterno: string;
    apellidoMaterno: string;
    correoElectronico: string;
    idColaborador: number;
    bolEnvioNotificaciones: number;
    bolUsuarioActivo: number;
    envioNotificaciones:string;
    usuarioActivo:string;
}